# open_api_docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- {}
paths:
...
  /api/circles:
    get:
      tags:
      - circle-controller
      operationId: index
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Circle"
    post:
      tags:
      - circle-controller
      operationId: createCircle
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Circle"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Circle"
  /api/circles/{circle}/notebooks:
    post:
      tags:
      - circle-controller
      operationId: createNotebookInCircle
      parameters:
      - name: circle
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteCreationDTO"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RedirectToNoteResponse"
  /api/circles/join:
    post:
      tags:
      - circle-controller
      operationId: joinCircle
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircleJoiningByInvitation"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Circle"
...
  /api/assessment/{assessmentQuestionInstance}/answer:
    post:
      tags:
      - assessment-controller
      operationId: answerQuestion
      parameters:
      - name: assessmentQuestionInstance
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerDTO"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentQuestionInstance"
  /api/assessment/{assessmentAttempt}:
    post:
      tags:
      - assessment-controller
      operationId: submitAssessmentResult
      parameters:
      - name: assessmentAttempt
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentAttempt"
  /api/assessment/questions/{notebook}:
    post:
      tags:
      - assessment-controller
      operationId: generateAssessmentQuestions
      parameters:
      - name: notebook
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AssessmentAttempt"
...
  /api/circles/{circle}:
    get:
      tags:
      - circle-controller
      operationId: showCircle
      parameters:
      - name: circle
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CircleForUserView"
...
  /api/assessment:
    get:
      tags:
      - assessment-controller
      operationId: getMyAssessments
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentAttempt"
...
components:
  schemas:
    Circle:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 255
          minLength: 1
      required:
      - id
      - name
    Ownership:
      type: object
      properties:
        id:
          type: integer
          format: int32
        circle:
          $ref: "#/components/schemas/Circle"
        ownerName:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        externalIdentifier:
          type: string
        ownership:
          $ref: "#/components/schemas/Ownership"
        dailyAssimilationCount:
          type: integer
          format: int32
        spaceIntervals:
          type: string
        admin:
          type: boolean
      required:
      - externalIdentifier
      - id
      - name
...
    MultipleChoicesQuestion:
      type: object
      properties:
        stem:
          type: string
        choices:
          type: array
          items:
            type: string
      required:
      - choices
      - stem
...
    SubscriptionDTO:
      type: object
      properties:
        dailyTargetOfNewNotes:
          type: integer
          format: int32
    Note:
      type: object
      properties:
        noteTopology:
          $ref: "#/components/schemas/NoteTopology"
        details:
          type: string
        parentId:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          readOnly: true
        wikidataId:
          type: string
      required:
      - createdAt
      - id
      - noteTopology
      - updatedAt
    NoteTopology:
      type: object
      properties:
        id:
          type: integer
          format: int32
        titleOrPredicate:
          type: string
        shortDetails:
          type: string
        linkType:
          type: string
          enum:
          - no link
          - related to
          - a specialization of
          - an application of
          - an instance of
          - a part of
          - tagged by
          - an attribute of
          - the opposite of
          - author of
          - using
          - an example of
          - before
          - similar to
          - confused with
        objectNoteTopology:
          $ref: "#/components/schemas/NoteTopology"
        parentOrSubjectNoteTopology:
          $ref: "#/components/schemas/NoteTopology"
        title:
          type: string
      required:
      - id
      - titleOrPredicate
    Notebook:
      type: object
      properties:
        id:
          type: integer
          format: int32
        certifiable:
          type: boolean
        notebookSettings:
          $ref: "#/components/schemas/NotebookSettings"
        creatorId:
          type: string
        title:
          type: string
        circle:
          $ref: "#/components/schemas/Circle"
        headNoteId:
          type: integer
          format: int32
        shortDetails:
          type: string
        updated_at:
          type: string
          format: date-time
      required:
      - headNoteId
      - id
      - notebookSettings
      - title
      - updated_at
    NotebookSettings:
      type: object
      properties:
        skipMemoryTrackingEntirely:
          type: boolean
        numberOfQuestionsInAssessment:
          type: integer
          format: int32
        certificateExpiry:
          type: string
          default: 1y
          example: 1y 6m
    Subscription:
      type: object
      properties:
        headNote:
          $ref: "#/components/schemas/Note"
        title:
          type: string
        id:
          type: integer
          format: int32
        dailyTargetOfNewNotes:
          type: integer
          format: int32
        user:
          $ref: "#/components/schemas/User"
        notebook:
          $ref: "#/components/schemas/Notebook"
        fromDTO:
          $ref: "#/components/schemas/SubscriptionDTO"
          writeOnly: true
      required:
      - id
...
    AnswerDTO:
      type: object
      properties:
        choiceIndex:
          type: integer
          format: int32
    Answer:
      type: object
      properties:
        id:
          type: integer
          format: int32
        choiceIndex:
          type: integer
          format: int32
        correct:
          type: boolean
      required:
      - correct
      - id
...
    NoteCreationDTO:
      type: object
      properties:
        newTitle:
          type: string
          maxLength: 150
          minLength: 1
        wikidataId:
          type: string
          pattern: ^$|Q\d+
      required:
      - newTitle
...
    RedirectToNoteResponse:
      type: object
      properties:
        noteId:
          type: integer
          format: int32
...
    AssessmentQuestionInstance:
      type: object
      properties:
        id:
          type: integer
          format: int32
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        answer:
          $ref: "#/components/schemas/Answer"
      required:
      - id
      - multipleChoicesQuestion
...
    CircleJoiningByInvitation:
      type: object
      properties:
        invitationCode:
          type: string
          maxLength: 20
          minLength: 10
      required:
      - invitationCode
...
    AssessmentAttempt:
      type: object
      properties:
        id:
          type: integer
          format: int32
        totalQuestionCount:
          type: integer
          format: int32
        answersCorrect:
          type: integer
          format: int32
        notebookId:
          type: integer
          format: int32
        notebookTitle:
          type: string
        submittedAt:
          type: string
          format: date-time
        isPass:
          type: boolean
        assessmentQuestionInstances:
          type: array
          items:
            $ref: "#/components/schemas/AssessmentQuestionInstance"
        certifiable:
          type: boolean
      required:
      - id
      - notebookId
...
    NotebooksViewedByUser:
      type: object
      properties:
        notebooks:
          type: array
          items:
            $ref: "#/components/schemas/Notebook"
        subscriptions:
          type: array
          items:
            $ref: "#/components/schemas/Subscription"
      required:
      - notebooks
...
    CircleForUserView:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        invitationCode:
          type: string
        notebooks:
          $ref: "#/components/schemas/NotebooksViewedByUser"
        members:
          type: array
          items:
            $ref: "#/components/schemas/UserForOtherUserView"
      required:
      - id
      - invitationCode
      - members
      - name
      - notebooks
    UserForOtherUserView:
      type: object
      properties:
        name:
          type: string
...